[None][refactor] scope Qwen multimodal AD helpers
