+++ tensorrt_llm/_torch/auto_deploy/models/custom/modeling_qwen3_5_moe.py
@@ -29,6 +29,7 @@ from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
 import torch
 import torch.nn.functional as F
+from PIL import Image
 from torch import nn
 from torch.export import Dim
 from transformers import AutoConfig
@@ -2313,6 +2314,83 @@ class Qwen3_5MoeADInputProcessor:
 
     def __getattr__(self, name: str):
         return getattr(self.base_processor, name)
+
+    @property
+    def get_num_multimodal_tokens(self):
+        """Delegate multimodal token counting to the wrapped Qwen HF processor."""
+        if hasattr(self.processor, "_get_num_multimodal_tokens"):
+            return self.processor._get_num_multimodal_tokens
+        raise NotImplementedError(
+            f"get_num_multimodal_tokens not implemented for {self.__class__.__name__}. "
+            "Please ensure the processor exposes _get_num_multimodal_tokens."
+        )
+
+    def get_num_tokens_per_image(self, *, image: Image.Image, **kwargs) -> int:
+        image_size = (image.height, image.width)
+        return self.get_num_multimodal_tokens([image_size], **kwargs)["num_image_tokens"][0]
+
+    def get_num_tokens_per_video(self, *, video: List[Image.Image], **kwargs) -> int:
+        video_size = (len(video), video[0].height, video[0].width)
+        num_video_tokens = self.get_num_multimodal_tokens(video_sizes=[video_size], **kwargs).get(
+            "num_video_tokens"
+        )
+        if num_video_tokens is None:
+            raise NotImplementedError("Underlying processor does not expose num_video_tokens.")
+        return num_video_tokens[0]
+
+    def get_vocab_size(self) -> Optional[int]:
+        """Return the tokenizer vocabulary size for Qwen multimodal hashing helpers."""
+        if self.tokenizer is not None and hasattr(self.tokenizer, "vocab_size"):
+            return int(self.tokenizer.vocab_size)
+        wrapped_tokenizer = getattr(self.tokenizer, "tokenizer", None)
+        if wrapped_tokenizer is not None and hasattr(wrapped_tokenizer, "vocab_size"):
+            return int(wrapped_tokenizer.vocab_size)
+        processor_tokenizer = getattr(self.processor, "tokenizer", None)
+        if processor_tokenizer is not None and hasattr(processor_tokenizer, "vocab_size"):
+            return int(processor_tokenizer.vocab_size)
+        return None
+
+    def get_mm_token_ids(self) -> Optional[torch.Tensor]:
+        if hasattr(self.processor, "mm_token_ids"):
+            return self.processor.mm_token_ids
+        sources = [
+            self.processor,
+            getattr(self.processor, "tokenizer", None),
+            self.tokenizer,
+            getattr(self.tokenizer, "tokenizer", None),
+        ]
+        token_ids = []
+        for source in sources:
+            if source is None:
+                continue
+            for attr in ("image_token_id", "video_token_id"):
+                value = getattr(source, attr, None)
+                if value is not None:
+                    token_ids.append(int(value))
+        if token_ids:
+            return torch.tensor(sorted(set(token_ids)), dtype=torch.int32)
+        return None
+
+    def get_mm_special_token_ids(self) -> Optional[torch.Tensor]:
+        if hasattr(self.processor, "mm_special_token_ids"):
+            return self.processor.mm_special_token_ids
+        sources = [
+            self.processor,
+            getattr(self.processor, "tokenizer", None),
+            self.tokenizer,
+            getattr(self.tokenizer, "tokenizer", None),
+        ]
+        token_ids = []
+        for source in sources:
+            if source is None:
+                continue
+            for attr in ("vision_start_token_id", "vision_end_token_id"):
+                value = getattr(source, attr, None)
+                if value is not None:
+                    token_ids.append(int(value))
+        if token_ids:
+            return torch.tensor(sorted(set(token_ids)), dtype=torch.int32)
+        return None
 
     def _build_multimodal_input(
         self,
