+++ tensorrt_llm/_torch/auto_deploy/shim/ad_executor.py
@@ -558,6 +558,104 @@ class ADEngine(ModelEngine):
             PyTorchModelEngine._execute_logit_post_processors, self
         )
 
+    def _store_prefill_multimodal_metadata(
+        self,
+        ordered_requests: RequestList,
+        input_pos: List[int],
+        cu_seqlen: List[int],
+        num_prefill_seqs: int,
+        extra_args: Dict[str, List[torch.Tensor]],
+    ) -> None:
+        """Stage per-request multimodal layout metadata needed by the VLM wrapper."""
+        if num_prefill_seqs == 0:
+            return
+
+        prefill_requests = ordered_requests[:num_prefill_seqs]
+        if not any(getattr(req, "multimodal_positions", None) for req in prefill_requests):
+            return
+
+        mm_item_cu_seqlen: List[int] = [0]
+        mm_item_types_flat: List[int] = []
+        mm_token_positions_flat: List[int] = []
+        mm_token_lengths_flat: List[int] = []
+        mm_special_offsets_cu_seqlen: List[int] = [0]
+        mm_special_offsets_flat: List[int] = []
+        flat_start_list: List[int] = []
+        count_list: List[int] = []
+        cumsum_total_mm = 0
+
+        for i, req in enumerate(prefill_requests):
+            begin_compute = input_pos[i]
+            end_compute = begin_compute + (cu_seqlen[i + 1] - cu_seqlen[i])
+            mm_pos = getattr(req, "multimodal_positions", None)
+            mm_len = getattr(req, "multimodal_lengths", None)
+            mm_item_types = (
+                req.py_multimodal_data.get("item_types", []) if req.py_multimodal_data else []
+            )
+            special_offsets = (
+                req.py_multimodal_data.get("special_token_offsets", [])
+                if req.py_multimodal_data
+                else []
+            )
+            mm_pos_list = list(mm_pos) if mm_pos is not None else []
+            mm_len_list = list(mm_len) if mm_len is not None else []
+            mm_item_cu_seqlen.append(mm_item_cu_seqlen[-1] + len(mm_pos_list))
+            mm_item_types_flat.extend(list(mm_item_types))
+            mm_token_positions_flat.extend(mm_pos_list)
+            mm_token_lengths_flat.extend(mm_len_list)
+            mm_special_offsets_cu_seqlen.append(
+                mm_special_offsets_cu_seqlen[-1] + len(special_offsets)
+            )
+            mm_special_offsets_flat.extend(list(special_offsets))
+            if not mm_pos or not mm_len:
+                flat_start_list.append(0)
+                count_list.append(0)
+                continue
+
+            runtime = MultimodalRuntimeData(
+                past_seen_token_num=begin_compute,
+                mm_token_positions=list(mm_pos),
+                mm_token_lengths=list(mm_len),
+                chunk_end_pos=end_compute,
+                special_token_offsets=list(special_offsets),
+            )
+            num_unseen = runtime.num_unseen_mm_tokens
+            num_in_chunk = runtime.num_mm_tokens_in_chunk
+            num_unseen_special = runtime.num_unseen_special_tokens
+            num_special_in_chunk = runtime.num_special_tokens_in_chunk
+            total_mm_i = runtime.total_mm_tokens_in_request
+            total_special_i = runtime.total_special_tokens_in_request
+            flat_start_i = cumsum_total_mm + num_unseen - num_unseen_special
+            flat_start_list.append(flat_start_i)
+            count_list.append(num_in_chunk - num_special_in_chunk)
+            cumsum_total_mm += total_mm_i - total_special_i
+
+        extra_args["mm_item_cu_seqlen"] = [
+            torch.tensor(mm_item_cu_seqlen, dtype=torch.int32, device="cpu")
+        ]
+        extra_args["mm_item_types"] = [
+            torch.tensor(mm_item_types_flat, dtype=torch.int32, device="cpu")
+        ]
+        extra_args["mm_token_positions"] = [
+            torch.tensor(mm_token_positions_flat, dtype=torch.int32, device="cpu")
+        ]
+        extra_args["mm_token_lengths"] = [
+            torch.tensor(mm_token_lengths_flat, dtype=torch.int32, device="cpu")
+        ]
+        extra_args["mm_special_offsets_cu_seqlen"] = [
+            torch.tensor(mm_special_offsets_cu_seqlen, dtype=torch.int32, device="cpu")
+        ]
+        extra_args["mm_special_offsets"] = [
+            torch.tensor(mm_special_offsets_flat, dtype=torch.int32, device="cpu")
+        ]
+        if self._enable_chunked_prefill:
+            extra_args["mm_chunk_flat_start"] = [
+                torch.tensor(flat_start_list, dtype=torch.int64, device="cpu")
+            ]
+            extra_args["mm_chunk_count"] = [
+                torch.tensor(count_list, dtype=torch.int64, device="cpu")
+            ]
+
     @nvtx_range("ad_prepare_inputs")
     def _prepare_inputs(
         self,
@@ -703,88 +801,13 @@ class ADEngine(ModelEngine):
         batch_info.extend([num_extend, num_extend_tokens])
         batch_info.extend([num_decode, num_decode_tokens])
 
-        # Stage per-request multimodal layout metadata so the VLM wrapper can reconstruct
-        # chunk-local mRoPE and compute request-level mRoPE deltas from full multimodal state.
-        if num_prefill_seqs > 0:
-            mm_item_cu_seqlen: List[int] = [0]
-            mm_item_types_flat: List[int] = []
-            mm_token_positions_flat: List[int] = []
-            mm_token_lengths_flat: List[int] = []
-            mm_special_offsets_cu_seqlen: List[int] = [0]
-            mm_special_offsets_flat: List[int] = []
-            flat_start_list: List[int] = []
-            count_list: List[int] = []
-            cumsum_total_mm = 0
-            for i in range(num_prefill_seqs):
-                req = ordered_requests[i]
-                begin_compute = input_pos[i]
-                end_compute = begin_compute + (cu_seqlen[i + 1] - cu_seqlen[i])
-                mm_pos = getattr(req, "multimodal_positions", None)
-                mm_len = getattr(req, "multimodal_lengths", None)
-                mm_item_types = (
-                    req.py_multimodal_data.get("item_types", []) if req.py_multimodal_data else []
-                )
-                special_offsets = (
-                    req.py_multimodal_data.get("special_token_offsets", [])
-                    if req.py_multimodal_data
-                    else []
-                )
-                mm_pos_list = list(mm_pos) if mm_pos is not None else []
-                mm_len_list = list(mm_len) if mm_len is not None else []
-                mm_item_cu_seqlen.append(mm_item_cu_seqlen[-1] + len(mm_pos_list))
-                mm_item_types_flat.extend(list(mm_item_types))
-                mm_token_positions_flat.extend(mm_pos_list)
-                mm_token_lengths_flat.extend(mm_len_list)
-                mm_special_offsets_cu_seqlen.append(
-                    mm_special_offsets_cu_seqlen[-1] + len(special_offsets)
-                )
-                mm_special_offsets_flat.extend(list(special_offsets))
-                if not mm_pos or not mm_len:
-                    flat_start_list.append(0)
-                    count_list.append(0)
-                    continue
-                runtime = MultimodalRuntimeData(
-                    past_seen_token_num=begin_compute,
-                    mm_token_positions=list(mm_pos),
-                    mm_token_lengths=list(mm_len),
-                    chunk_end_pos=end_compute,
-                    special_token_offsets=list(special_offsets),
-                )
-                num_unseen = runtime.num_unseen_mm_tokens
-                num_in_chunk = runtime.num_mm_tokens_in_chunk
-                num_unseen_special = runtime.num_unseen_special_tokens
-                num_special_in_chunk = runtime.num_special_tokens_in_chunk
-                total_mm_i = runtime.total_mm_tokens_in_request
-                total_special_i = runtime.total_special_tokens_in_request
-                flat_start_i = cumsum_total_mm + num_unseen - num_unseen_special
-                flat_start_list.append(flat_start_i)
-                count_list.append(num_in_chunk - num_special_in_chunk)
-                cumsum_total_mm += total_mm_i - total_special_i
-            extra_args["mm_item_cu_seqlen"] = [
-                torch.tensor(mm_item_cu_seqlen, dtype=torch.int32, device="cpu")
-            ]
-            extra_args["mm_item_types"] = [
-                torch.tensor(mm_item_types_flat, dtype=torch.int32, device="cpu")
-            ]
-            extra_args["mm_token_positions"] = [
-                torch.tensor(mm_token_positions_flat, dtype=torch.int32, device="cpu")
-            ]
-            extra_args["mm_token_lengths"] = [
-                torch.tensor(mm_token_lengths_flat, dtype=torch.int32, device="cpu")
-            ]
-            extra_args["mm_special_offsets_cu_seqlen"] = [
-                torch.tensor(mm_special_offsets_cu_seqlen, dtype=torch.int32, device="cpu")
-            ]
-            extra_args["mm_special_offsets"] = [
-                torch.tensor(mm_special_offsets_flat, dtype=torch.int32, device="cpu")
-            ]
-            if self._enable_chunked_prefill:
-                extra_args["mm_chunk_flat_start"] = [
-                    torch.tensor(flat_start_list, dtype=torch.int64, device="cpu")
-                ]
-                extra_args["mm_chunk_count"] = [
-                    torch.tensor(count_list, dtype=torch.int64, device="cpu")
-                ]
+        self._store_prefill_multimodal_metadata(
+            ordered_requests=ordered_requests,
+            input_pos=input_pos,
+            cu_seqlen=cu_seqlen,
+            num_prefill_seqs=num_prefill_seqs,
+            extra_args=extra_args,
+        )
 
         # update the sequence info object now (also triggers rescatter + host_prepare internally)
         self.cache_seq_interface.info.nest_sequences(
